requirements.txt
fastapi
uvicorn
openai
chromadb
sentence-transformers
numpy
pillow
python-dotenv
